feat: removed all node command add missing cleanups
