--- cmd/vega/node/node_pre_windows.go
@@ -1,9 +1,0 @@
-//go:build windows
-// +build windows
-
-package node
-
-// SetUlimits is currently a no-op on Windows.
-func (l *NodeCommand) SetUlimits() error {
-	return nil
-}
